+++ tests/test_util.py
@@ -1,12 +1,16 @@
+import logging
 from time import sleep
 import pytest
-from mpservice.util import Thread, TimeoutError, Process
+from mpservice.util import Thread, TimeoutError, SpawnProcess
+from mpservice.remote_exception import RemoteException
+
+
+logger = logging.getLogger(__name__)
 
 
 def delay_double(x, delay=2):
     sleep(delay)
     if x < 10:
-        print('returning', x * 2)
         return x * 2
     raise ValueError(x)
 
@@ -14,6 +18,7 @@ def delay_double(x, delay=2):
 def _test_thread_process(cls):
     t = cls(target=delay_double, args=(3,))
     t.start()
+    logger.info('to sleep')
     sleep(0.1)
     assert not t.done()
     assert t.is_alive()
@@ -25,7 +30,7 @@ def _test_thread_process(cls):
     assert t.exception() is None
     t.join()
 
-    t = Thread(target=delay_double, args=(12,))
+    t = cls(target=delay_double, args=(12,))
     t.start()
     with pytest.raises(TimeoutError):
         y = t.result(0.2)
@@ -35,6 +40,7 @@ def _test_thread_process(cls):
 
     e = t.exception()
     assert type(e) is ValueError
+
     t.join()
 
 
@@ -48,7 +54,4 @@ def _process_func(x):
 
 
 def test_process():
-    p = Process(target=_process_func, args=(123,))
-    p.start()
-    p.join()
-    print('result:', p.result())
+    _test_thread_process(SpawnProcess)